Added 2 parameters and 1 branch
Added 2 additional parameters. run1 is the number label of the first data file being analysed. runs is the number of data files that will be analysed. For example, $ root calc_rms_run.C'(1165004,3)' will analyse 3 data files labelled 1165003, 1165004, 1165005 consecutively. If any of them don't exist, it will be skipped. For example, using the above code, if 1165004 doesn't exist, only 1165003 and 1165005 will be analysed.

The 3rd parameter is the directory of the data files. By default it is (char*)"/scratch3/wparker2/dmtpc2/data/2017/06/raw".

User can choose to store or don't store the intensity values of all pixels when the code runs. The size of output file will be inflated by more than 1000 times if the user choose to store these values.